Break into select rendering
