+++ src/components/interactive-builder/modals/question/question-form/question-types/inputs/obs/obs-type-question.component.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, useMemo, useRef, useState } from 'react';
 import { FormLabel, InlineNotification, MultiSelect, FormGroup, Tag, Stack } from '@carbon/react';
 import { useTranslation } from 'react-i18next';
-import ConceptSearch from './concept-search.component';
+import ConceptSearch from '../../../common/concept-search/concept-search.component';
 import { useFormField } from '../../../../form-field-context';
 import type { Concept, ConceptMapping, DatePickerType } from '@types';
 import styles from './obs-type-question.scss';
@@ -13,24 +13,8 @@ interface AnswerItem {
 
 const ObsTypeQuestion: React.FC = () => {
   const { t } = useTranslation();
-  const { formField, setFormField } = useFormField();
-  const [selectedConcept, setSelectedConcept] = useState<Concept>(null);
-  const [selectedAnswers, setSelectedAnswers] = useState<Array<AnswerItem>>(
-    formField.questionOptions?.answers
-      ? formField.questionOptions.answers.map((answer) => ({
-          id: answer.concept,
-          text: answer.label,
-        }))
-      : [],
-  );
-  const [addedAnswers, setAddedAnswers] = useState<AnswerItem[]>([]);
+  const { formField, setFormField, concept, setConcept } = useFormField();
   const [conceptMappings, setConceptMappings] = useState<Array<ConceptMapping>>([]);
-  const initiallySelectedAnswers = useRef(
-    formField?.questionOptions?.answers?.map((answer) => ({
-      id: answer.concept,
-      text: answer.label,
-    })),
-  );
 
   const getDatePickerType = useCallback((concept: Concept): DatePickerType | null => {
     const conceptDataType = concept.datatype.name;
@@ -47,20 +31,20 @@ const ObsTypeQuestion: React.FC = () => {
   }, []);
 
   const handleConceptSelect = useCallback(
-    (concept: Concept) => {
-      setSelectedConcept(concept);
-      if (concept) {
-        const datePickerType = getDatePickerType(concept);
+    (selectedConcept: Concept) => {
+      setConcept(selectedConcept);
+      if (selectedConcept) {
+        const datePickerType = getDatePickerType(selectedConcept);
         setFormField((prevField) => ({
           ...prevField,
           questionOptions: {
             ...prevField.questionOptions,
-            concept: concept.uuid,
+            concept: selectedConcept.uuid,
           },
           ...(datePickerType && { datePickerFormat: datePickerType }),
         }));
         setConceptMappings(
-          concept?.mappings?.map((conceptMapping) => {
+          selectedConcept?.mappings?.map((conceptMapping) => {
             const data = conceptMapping.display.split(': ');
             return {
               relationship: conceptMapping.conceptMapType.display,
@@ -71,11 +55,11 @@ const ObsTypeQuestion: React.FC = () => {
         );
       }
     },
-    [getDatePickerType, setFormField],
+    [getDatePickerType, setFormField, setConcept],
   );
 
   const clearSelectedConcept = useCallback(() => {
-    setSelectedConcept(null);
+    setConcept(null);
     setConceptMappings([]);
 
     setFormField((prevFormField) => {
@@ -89,104 +73,7 @@ const ObsTypeQuestion: React.FC = () => {
       }
       return updatedFormField;
     });
-  }, [setFormField]);
-
-  const selectAnswers = useCallback(
-    ({ selectedItems }: { selectedItems: Array<AnswerItem> }) => {
-      const mappedAnswers = selectedItems.map((answer) => ({
-        concept: answer.id,
-        label: answer.text,
-      }));
-
-      setSelectedAnswers(selectedItems);
-
-      setFormField((prevField) => {
-        const currentAnswers = prevField.questionOptions?.answers || [];
-        if (JSON.stringify(currentAnswers) === JSON.stringify(mappedAnswers)) {
-          return prevField;
-        }
-        return {
-          ...prevField,
-          questionOptions: {
-            ...prevField.questionOptions,
-            answers: mappedAnswers,
-          },
-        };
-      });
-    },
-    [setFormField],
-  );
-
-  const handleDeleteAdditionalAnswer = useCallback(
-    (id: string) => {
-      setAddedAnswers((prevAnswers) => prevAnswers.filter((answer) => answer.id !== id));
-      setFormField((prevFormField) => {
-        const selectedAnswers = prevFormField.questionOptions?.answers ?? [];
-        return {
-          ...prevFormField,
-          questionOptions: {
-            ...prevFormField.questionOptions,
-            answers: selectedAnswers.filter((answer) => answer.concept !== id),
-          },
-        };
-      });
-    },
-    [setFormField],
-  );
-
-  const handleSelectAdditionalAnswer = useCallback(
-    (concept: Concept) => {
-      const newAnswer = { id: concept.uuid, text: concept.display };
-      const answerExistsInSelected = selectedAnswers.some((answer) => answer.id === newAnswer.id);
-      const answerExistsInAdded = addedAnswers.some((answer) => answer.id === newAnswer.id);
-      if (!answerExistsInSelected && !answerExistsInAdded) {
-        setAddedAnswers((prevAnswers) => [...prevAnswers, newAnswer]);
-        setFormField((prevFormField) => {
-          const existingAnswers = prevFormField.questionOptions?.answers ?? [];
-          existingAnswers.push({ concept: concept.uuid, label: concept.display });
-          return {
-            ...prevFormField,
-            questionOptions: {
-              ...prevFormField.questionOptions,
-              answers: existingAnswers,
-            },
-          };
-        });
-      }
-    },
-    [selectedAnswers, addedAnswers, setFormField],
-  );
-
-  const answerItems = useMemo(() => {
-    // Convert answers from the concept to items format
-    const conceptAnswerItems =
-      selectedConcept?.answers?.map((answer) => ({
-        id: answer.uuid,
-        text: answer.display,
-      })) ?? [];
-
-    const formFieldAnswers = formField.questionOptions?.answers ?? [];
-
-    // If no answers from concept but we have form field answers, use those
-    if (conceptAnswerItems.length === 0 && formFieldAnswers.length > 0) {
-      return formFieldAnswers.map((answer) => ({
-        id: answer.concept,
-        text: answer.label,
-      }));
-    }
-
-    // Merge concept answers with any additional form field answers
-    const additionalAnswers = formFieldAnswers
-      .filter((answer) => !conceptAnswerItems.some((item) => item.id === answer.concept))
-      .map((answer) => ({
-        id: answer.concept,
-        text: answer.label,
-      }));
-
-    return [...conceptAnswerItems, ...additionalAnswers];
-  }, [selectedConcept?.answers, formField.questionOptions?.answers]);
-
-  const convertAnswerItemsToString = useCallback((item: AnswerItem) => item.text, []);
+  }, [setFormField, setConcept]);
 
   return (
     <Stack gap={5}>
@@ -196,7 +83,7 @@ const ObsTypeQuestion: React.FC = () => {
         onSelectConcept={handleConceptSelect}
       />
 
-      {selectedConcept?.allowDecimal === false && (
+      {concept?.allowDecimal === false && (
         <InlineNotification
           kind="info"
           lowContrast
@@ -227,55 +114,6 @@ const ObsTypeQuestion: React.FC = () => {
           </table>
         </FormGroup>
       ) : null}
-
-      {answerItems.length > 0 && (
-        <MultiSelect
-          className={styles.multiSelect}
-          direction="top"
-          id="selectAnswers"
-          items={answerItems}
-          itemToString={convertAnswerItemsToString}
-          onChange={selectAnswers}
-          size="md"
-          selectedItems={selectedAnswers}
-          initialSelectedItems={initiallySelectedAnswers}
-          titleText={t('selectAnswersToDisplay', 'Select answers to display')}
-        />
-      )}
-
-      {selectedAnswers.length > 0 && (
-        <div>
-          {selectedAnswers.map((answer) => (
-            <Tag className={styles.tag} key={answer.id} type={'blue'}>
-              {answer.text}
-            </Tag>
-          ))}
-        </div>
-      )}
-
-      {selectedConcept && selectedConcept.datatype?.name === 'Coded' && (
-        <>
-          <ConceptSearch
-            label={t('searchForAnswerConcept', 'Search for a concept to add as an answer')}
-            onSelectConcept={handleSelectAdditionalAnswer}
-          />
-          {addedAnswers.length > 0 ? (
-            <div>
-              {addedAnswers.map((answer) => (
-                <Tag className={styles.tag} key={answer.id} type={'blue'}>
-                  {answer.text}
-                  <button
-                    className={styles.conceptAnswerButton}
-                    onClick={() => handleDeleteAdditionalAnswer(answer.id)}
-                  >
-                    X
-                  </button>
-                </Tag>
-              ))}
-            </div>
-          ) : null}{' '}
-        </>
-      )}
     </Stack>
   );
 };
